+++ rust/operator-binary/src/zk_controller.rs
@@ -15,6 +15,10 @@ use crate::{
 };
 use fnv::FnvHasher;
 use snafu::{OptionExt, ResultExt, Snafu};
+use stackable_operator::k8s_openapi::api::{
+    core::v1::ServiceAccount,
+    rbac::v1::{RoleRef, Subject},
+};
 use stackable_operator::{
     builder::{
         ConfigMapBuilder, ContainerBuilder, ObjectMetaBuilder, PodBuilder,
@@ -44,6 +48,7 @@ use stackable_operator::{
     product_config_utils::{transform_all_roles_to_config, validate_all_roles_and_groups_config},
     role_utils::RoleGroupRef,
 };
+use stackable_operator::{k8s_openapi::api::rbac::v1::RoleBinding, kube::ResourceExt};
 use stackable_zookeeper_crd::{
     ZookeeperCluster, ZookeeperClusterStatus, ZookeeperConfig, ZookeeperRole, CLIENT_TLS_DIR,
     QUORUM_TLS_DIR, STACKABLE_CONFIG_DIR, STACKABLE_DATA_DIR, STACKABLE_RW_CONFIG_DIR,
@@ -51,6 +56,8 @@ use stackable_zookeeper_crd::{
 use strum::{EnumDiscriminants, IntoStaticStr};
 
 const FIELD_MANAGER_SCOPE: &str = "zookeepercluster";
+
+const SERVICE_ACCOUNT: &str = "zookeeper-serviceaccount";
 
 pub struct Ctx {
     pub client: stackable_operator::client::Client,
@@ -143,6 +150,14 @@ pub enum Error {
     InvalidJavaHeapConfig {
         source: stackable_operator::error::Error,
     },
+    #[snafu(display("failed to create RBAC service account: {source}"))]
+    ApplyServiceAccount {
+        source: stackable_operator::error::Error,
+    },
+    #[snafu(display("failed to create RBAC role binding: {source}"))]
+    ApplyRoleBinding {
+        source: stackable_operator::error::Error,
+    },
 }
 type Result<T, E = Error> = std::result::Result<T, E>;
 
@@ -178,6 +193,8 @@ impl ReconcilerError for Error {
                 ..
             } => Some(authentication_class.clone().erase()),
             Error::InvalidJavaHeapConfig { .. } => None,
+            Error::ApplyServiceAccount { .. } => None,
+            Error::ApplyRoleBinding { .. } => None,
         }
     }
 }
@@ -232,6 +249,16 @@ pub async fn reconcile_zk(
     } else {
         None
     };
+
+    let (rbac_sa, rbac_rolebinding) = build_zk_rbac_resources(&zk)?;
+    client
+        .apply_patch(FIELD_MANAGER_SCOPE, &rbac_sa, &rbac_sa)
+        .await
+        .with_context(|_| ApplyServiceAccountSnafu)?;
+    client
+        .apply_patch(FIELD_MANAGER_SCOPE, &rbac_rolebinding, &rbac_rolebinding)
+        .await
+        .with_context(|_| ApplyRoleBindingSnafu)?;
 
     let server_role_service = build_server_role_service(&zk)?;
     let server_role_service = client
@@ -300,6 +327,38 @@ pub async fn reconcile_zk(
         .context(ApplyStatusSnafu)?;
 
     Ok(controller::Action::await_change())
+}
+
+pub fn build_zk_rbac_resources(zk: &ZookeeperCluster) -> Result<(ServiceAccount, RoleBinding)> {
+    let service_account = ServiceAccount {
+        metadata: ObjectMetaBuilder::new()
+            .name_and_namespace(zk)
+            .name(SERVICE_ACCOUNT.to_string())
+            .with_label("managed-by".to_string(), "zookeeper-operator".to_string())
+            .build(),
+        ..ServiceAccount::default()
+    };
+
+    let role_binding = RoleBinding {
+        metadata: ObjectMetaBuilder::new()
+            .name_and_namespace(zk)
+            .name("zookeeper-rolebinding".to_string())
+            .with_label("managed-by".to_string(), "zookeeper-operator".to_string())
+            .build(),
+        role_ref: RoleRef {
+            kind: "ClusterRole".to_string(),
+            name: "zookeeper-clusterrole".to_string(),
+            api_group: "rbac.authorization.k8s.io".to_string(),
+        },
+        subjects: Some(vec![Subject {
+            kind: "ServiceAccount".to_string(),
+            name: SERVICE_ACCOUNT.to_string(),
+            namespace: zk.namespace(),
+            ..Subject::default()
+        }]),
+    };
+
+    Ok((service_account, role_binding))
 }
 
 /// The server-role service is the primary endpoint that should be used by clients that do not perform internal load balancing,
@@ -596,7 +655,8 @@ fn build_server_rolegroup_statefulset(
         .security_context(PodSecurityContext {
             fs_group: Some(1000),
             ..PodSecurityContext::default()
-        });
+        })
+        .service_account_name(SERVICE_ACCOUNT);
 
     if zk.is_client_secure() {
         let secret_class = if let Some(auth_class) = client_authentication_class {